Move validation back to common
