+++ microprofile/testing/common/src/main/java/io/helidon/microprofile/testing/common/TestValidator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.helidon.microprofile.testing.validator;
+package io.helidon.microprofile.testing.common;
 
 /**
  * Validator for test class.
