+++ microprofile/testing/testng/pom.xml
@@ -56,10 +56,6 @@
             <groupId>io.helidon.microprofile.testing</groupId>
             <artifactId>helidon-microprofile-testing-jaxrs</artifactId>
         </dependency>
-        <dependency>
-            <groupId>io.helidon.microprofile.testing</groupId>
-            <artifactId>helidon-microprofile-testing-validator</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
