+++ microprofile/testing/testng/src/main/java/io/helidon/microprofile/testing/testng/TestNgJaxRsValidator.java
@@ -15,7 +15,7 @@
  */
 package io.helidon.microprofile.testing.testng;
 
-import io.helidon.microprofile.testing.validator.JaxRsValidator;
+import io.helidon.microprofile.testing.common.JaxRsValidator;
 
 /**
  * An implementation for JaxRs Validator for Junit.
